+++ loadui-project/loadui-api/src/main/java/com/eviware/loadui/api/statistics/store/Entry.java
@@ -31,7 +31,7 @@ public interface Entry
 	 * 
 	 * @return
 	 */
-	public int getTimestamp();
+	public long getTimestamp();
 
 	/**
 	 * Gets the names of the values in the Entry.
